[Feature #18595] Alias String#-@ as String#dedup
